--- client-side/newfile.txt
@@ -1,1 +1,0 @@
-start frontend
