+++ client-side/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "myapp",
+  "name": "client-side",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
