+++ client-side/src/App.css
@@ -1,3 +1,8 @@
+* {
+  box-sizing: border-box;
+  margin: 0px;
+  padding: 0px;
+}
 .App {
   text-align: center;
 }
