+++ client-side/src/App.js
@@ -1,0 +1,7 @@
+import "./App.css";
+
+function App() {
+  return <div className="App">Hello</div>;
+}
+
+export default App;
